+++ core/src/main/java/tc/oc/pgm/command/MapDevCommand.java
@@ -3,18 +3,23 @@ package tc.oc.pgm.command;
 import static net.kyori.adventure.text.Component.join;
 import static net.kyori.adventure.text.Component.text;
 import static tc.oc.pgm.command.util.ParserConstants.CURRENT;
+import static tc.oc.pgm.util.bukkit.Effects.EFFECTS;
 import static tc.oc.pgm.util.text.TextException.exception;
 
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.JoinConfiguration;
 import net.kyori.adventure.text.format.NamedTextColor;
+import org.bukkit.Color;
+import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 import org.bukkit.util.Vector;
 import org.incendo.cloud.annotations.Argument;
 import org.incendo.cloud.annotations.Command;
@@ -31,7 +36,7 @@ import tc.oc.pgm.util.Audience;
 import tc.oc.pgm.util.PrettyPaginatedComponentResults;
 import tc.oc.pgm.util.block.BlockFaces;
 import tc.oc.pgm.util.block.BlockVectors;
-import tc.oc.pgm.util.bukkit.Effects;
+import tc.oc.pgm.util.material.BlockMaterialData;
 import tc.oc.pgm.util.material.MaterialData;
 import tc.oc.pgm.util.text.TextFormatter;
 import tc.oc.pgm.variables.Variable;
@@ -131,31 +136,103 @@ public class MapDevCommand {
     if (!reg.getBounds().isBlockFinite()) {
       throw exception("Region is not finite");
     }
+    new DisplayRunner(pl, reg);
+  }
 
-    var block = MaterialData.block(Material.GLASS);
+  private static class DisplayRunner implements Runnable {
+    private static final BlockMaterialData GLASS = MaterialData.block(Material.GLASS);
+    private static final float SPREAD_MUL = 0.2f;
+    private static final float COUNT_MUL = 0.1f;
 
-    PGM.get().getAsyncExecutor().execute(() -> reg.getBlockVectors().forEach(bv -> {
-      var b = BlockVectors.blockAt(world, bv);
-      for (var dir : BlockFaces.NEIGHBORS) {
-        if (!reg.contains(b.getRelative(dir))) {
-          block.sendBlockChange(pl, b.getLocation());
-          break;
+    private final Player player;
+    private final Region.Static region;
+    private final Location loc;
+
+    private final float[][] directions;
+    private final double[][] edges, vertices;
+
+    private final Future<?> task;
+    private int ticks = 150;
+
+    private DisplayRunner(Player player, Region.Static region) {
+      this.player = player;
+      this.region = region;
+      this.loc = new Location(player.getWorld(), 0, 0, 0);
+
+      var bounds = region.getBounds();
+      Vector min = bounds.getMin(), max = bounds.getMax();
+      boolean hasBottom = min.getY() >= -64, hasTop = max.getY() <= 320;
+      if (!hasBottom) min.setY(-32);
+      if (!hasTop) max.setY(288);
+
+      var center = min.clone().add(max).multiply(0.5);
+      var size = max.clone().subtract(min);
+
+      this.directions = new float[][] {
+        {(float) size.getX() * SPREAD_MUL, 0, 0, (float) Math.max(10, size.getX() * COUNT_MUL)},
+        {0, (float) size.getY() * SPREAD_MUL, 0, (float) Math.max(10, size.getY() * COUNT_MUL)},
+        {0, 0, (float) size.getZ() * SPREAD_MUL, (float) Math.max(10, size.getZ() * COUNT_MUL)}
+      };
+      if (!hasBottom) min.setY(Double.NaN);
+      if (!hasTop) max.setY(Double.NaN);
+
+      this.edges = new double[][] {
+        {center.getX(), min.getY(), min.getZ()},
+        {center.getX(), min.getY(), max.getZ()},
+        {center.getX(), max.getY(), min.getZ()},
+        {center.getX(), max.getY(), max.getZ()},
+        {min.getX(), center.getY(), min.getZ()},
+        {min.getX(), center.getY(), max.getZ()},
+        {max.getX(), center.getY(), min.getZ()},
+        {max.getX(), center.getY(), max.getZ()},
+        {min.getX(), min.getY(), center.getZ()},
+        {min.getX(), max.getY(), center.getZ()},
+        {max.getX(), min.getY(), center.getZ()},
+        {max.getX(), max.getY(), center.getZ()},
+      };
+      this.vertices = new double[][] {
+        {min.getX(), min.getY(), min.getZ()},
+        {min.getX(), min.getY(), max.getZ()},
+        {min.getX(), max.getY(), min.getZ()},
+        {min.getX(), max.getY(), max.getZ()},
+        {max.getX(), min.getY(), min.getZ()},
+        {max.getX(), min.getY(), max.getZ()},
+        {max.getX(), max.getY(), min.getZ()},
+        {max.getX(), max.getY(), max.getZ()},
+      };
+
+      region.getBlockVectors().forEach(bv -> {
+        var b = BlockVectors.blockAt(loc.getWorld(), bv);
+        for (var dir : BlockFaces.NEIGHBORS) {
+          if (!region.contains(b.getRelative(dir))) {
+            GLASS.sendBlockChange(player, b.getLocation());
+            break;
+          }
         }
+      });
+
+      task = PGM.get().getExecutor().scheduleWithFixedDelay(this, 0, 100, TimeUnit.MILLISECONDS);
+    }
+
+    @Override
+    public void run() {
+      if (ticks-- < 0) {
+        task.cancel(true);
+        region.getBlocks(player.getWorld()).forEach(b -> MaterialData.block(b)
+            .sendBlockChange(player, b.getLocation()));
+        return;
       }
-    }));
-
-    PGM.get()
-        .getAsyncExecutor()
-        .schedule(
-            () -> reg.getBlockVectors().forEach(bv -> {
-              var b = BlockVectors.blockAt(world, bv);
-              MaterialData.block(b).sendBlockChange(pl, b.getLocation());
-            }),
-            15,
-            TimeUnit.SECONDS);
-
-    var bounds = region.getStatic(pl.getWorld()).getBounds();
-    Vector min = bounds.getMin(), max = bounds.getMax();
-    Effects.EFFECTS.renderRegion(pl, min, max, PGM.get().getExecutor());
+      for (int i = 0; i < edges.length; i++) {
+        if (Double.isNaN(edges[i][1])) continue;
+        loc.set(edges[i][0], edges[i][1], edges[i][2]);
+        var dirs = directions[(i / 4)];
+        EFFECTS.spawnFlame(player, loc, dirs[0], dirs[1], dirs[2], (int) dirs[3]);
+      }
+      for (double[] vertex : vertices) {
+        if (Double.isNaN(vertex[1])) continue;
+        loc.set(vertex[0], vertex[1], vertex[2]);
+        EFFECTS.coloredDust(player, loc, Color.RED);
+      }
+    }
   }
 }
