+++ platform/platform-modern/src/main/java/tc/oc/pgm/platform/modern/impl/ModernEffects.java
@@ -3,8 +3,6 @@ package tc.oc.pgm.platform.modern.impl;
 import static tc.oc.pgm.util.material.ColorUtils.COLOR_UTILS;
 import static tc.oc.pgm.util.platform.Supports.Variant.PAPER;
 
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 import org.bukkit.Color;
 import org.bukkit.DyeColor;
 import org.bukkit.Effect;
@@ -12,12 +10,8 @@ import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Particle;
 import org.bukkit.World;
-import org.bukkit.entity.BlockDisplay;
 import org.bukkit.entity.Player;
-import org.bukkit.util.Vector;
-import org.joml.Matrix4f;
 import tc.oc.pgm.platform.modern.material.ModernBlockMaterialData;
-import tc.oc.pgm.util.bukkit.BukkitUtils;
 import tc.oc.pgm.util.bukkit.Effects;
 import tc.oc.pgm.util.material.BlockMaterialData;
 import tc.oc.pgm.util.platform.Supports;
@@ -87,22 +81,7 @@ public class ModernEffects implements Effects {
   }
 
   @Override
-  public void renderRegion(
-      Player player, Vector min, Vector max, ScheduledExecutorService executor) {
-    var center = min.toLocation(player.getWorld());
-    var size = max.clone().subtract(min);
-    var entity = player.getWorld().spawn(center, BlockDisplay.class, e -> {
-      e.setVisibleByDefault(false);
-      e.setPersistent(false);
-
-      e.setBlock(Material.WHITE_STAINED_GLASS.createBlockData());
-      e.setShadowRadius(500);
-      e.setShadowRadius(0);
-      e.setTransformationMatrix(
-          new Matrix4f().scale((float) size.getX(), (float) size.getY(), (float) size.getZ()));
-    });
-
-    player.showEntity(BukkitUtils.getPlugin(), entity);
-    executor.schedule(entity::remove, 15, TimeUnit.SECONDS);
+  public void spawnFlame(Player player, Location loc, float x, float y, float z, int amt) {
+    player.spawnParticle(Particle.FLAME, loc, amt, x, y, z, 0, null, true);
   }
 }
